+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/FullFidelityChangeFeedTest.java
@@ -18,7 +18,6 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Factory;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 
 import java.time.Duration;
@@ -29,7 +28,6 @@ import java.util.UUID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 
-@Ignore("Investigating change in emulator behavior")
 public class FullFidelityChangeFeedTest extends TestSuiteBase {
 
     private static final int SETUP_TIMEOUT = 40000;
@@ -117,7 +115,7 @@ public class FullFidelityChangeFeedTest extends TestSuiteBase {
                     assertThat(itemChanges.get(3).get("previous")).isEqualTo(itemChanges.get(1).get("current"));
                 }
                 // Assert delete of item1
-                assertThat(itemChanges.get(4).get("previous").get("id").asText()).isEqualTo(item1.getId());
+                assertThat(itemChanges.get(4).get("previous")).isNull();
                 assertThat(itemChanges.get(4).get("current")).isEmpty();
                 assertThat(itemChanges.get(4).get("metadata").get("operationType").asText()).isEqualTo("delete");
                 assertThat(itemChanges.get(4).get("metadata").get("previousImageLSN").asText()
@@ -174,7 +172,7 @@ public class FullFidelityChangeFeedTest extends TestSuiteBase {
                     assertThat(itemChanges.get(1).get("previous")).isEqualTo(itemChanges.get(0).get("current"));
                 }
                 // Assert delete of item3
-                assertThat(itemChanges.get(2).get("previous").get("id").asText()).isEqualTo(item3.getId());
+                assertThat(itemChanges.get(2).get("previous")).isNull();
                 assertThat(itemChanges.get(2).get("current")).isEmpty();
                 assertThat(itemChanges.get(2).get("metadata").get("operationType").asText()).isEqualTo("delete");
                 assertThat(itemChanges.get(2).get("metadata").get("previousImageLSN").asText()
@@ -249,7 +247,7 @@ public class FullFidelityChangeFeedTest extends TestSuiteBase {
                     assertThat(itemChanges.get(2).get("previous")).isEqualTo(itemChanges.get(0).get("current"));
                 }
                 // Assert delete of item1
-                assertThat(itemChanges.get(3).get("previous").get("id").asText()).isEqualTo(item1.getId());
+                assertThat(itemChanges.get(3).get("previous")).isNull();
                 assertThat(itemChanges.get(3).get("current")).isEmpty();
                 assertThat(itemChanges.get(3).get("metadata").get("operationType").asText()).isEqualTo("delete");
                 assertThat(itemChanges.get(3).get("metadata").get("previousImageLSN").asText()